dynamic: change the way io_is_busy works
when set to 0 or 1 it works as a flag,
when set to 2..99, it is an upper bound of what percentage
of total cpu time that can be accounted as busy due to i/o.

i.e. if io_is_busy is 15, and io_wait is 7%, it'll be added entirely as
busy, but is io_wait jumps to 33%, only 15% of cpu time will be
considered busy due to i/o. (it's not 15% * 33% !!)